services/iot-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
influxdb-client==1.39.0
redis==5.0.1
paho-mqtt==1.6.1
websockets==12.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
celery==5.3.4 